Code Cleanup, Editor Config alignment

+++ Project/EngineManaged/Source/World/Terrain.cs
@@ -1,4 +1,4 @@
 namespace SlimeCore.Core.World;
 
-public enum Terrain : byte { Grass, Rock, Water, Lava, Mud, Ice, Speed }
+public enum Terrain : int { Grass, Rock, Water, Lava, Mud, Ice, Speed }
